Some nice start menu frontend. Plus template system works in other modules.
(Closes #21)
The ability for other modules to have their own templates is all set,
and I'm quite happy with it (closes #23). Now I can start making the
actual project creation system. The template currently set up for C#
is very basic and definitely not the final product. The frontend code
is more complex than I initially expected, but I'm proud of it.

I realized the new ArgumentBase system (basically a full rewrite of the
one I used in the last iteration of this project) doesn't have support
for enums, I missed that one. So I've made issue #41 for it. I'll get
to it sometime soon.