Added a printhelper method for text and made argumentbase parse enums.
The idea of the PrintDescription method turned into the PrintLargeString
method, so it could be generalized a little bit. Still not sure how I
feel about the scaling, might tweak that later. Like with the loading
bar, it looks kind of ugly at large terminal sizes. Still, closes #42.

Also, turns out enum parsing was not nearly as difficult as I thought it
would be. I didn't even have to use reflection, so in a way it's easier
than regular IParsable parsing. Closes #41.