Plenty of new files. Reading project information works now.
I really don't know if I should have made so many issues for this. But
it was a lot of files, all being said. The ProjectContext class can now
read the project file in a modular way, where the properties are
dynamically loaded based on the compiler and output variants. I'm also
handing exceptions better than I was in the last iteration. Something I
might want to do before I call it quits on this subject is make a new
exception type that contains the XML context (see issue #52).

Anyway, this closes a lot of issues. Closes #51, closes #50, closes #49,
closes #46, closes #45, closes #44, which closes the big one, #43. Again,
not quite sure this commit deserves so many issues, oops.